requirements.txt
Flask
pillow
pandas
numpy
scikit-learn